+++ pol-core/pol/decay.h
@@ -11,5 +11,5 @@ Notes
 void decay_items();
 void decay_single_zone();
 
-extern int cycles_per_decay_worldzone;
-extern int cycles_until_decay_worldzone;
+extern size_t cycles_per_decay_worldzone;
+extern size_t cycles_until_decay_worldzone;
