+++ pol-core/pol/network/client.cpp
@@ -263,13 +263,13 @@ void Client::itemizeclientversion(const std::string& ver, VersionDetailStruct& d
 {
 	try
 	{
-		int dot1 = ver.find_first_of('.',0);
-		int dot2 = ver.find_first_of('.',dot1 + 1);
-		int dot3 = ver.find_first_of('.',dot2 + 1);
+		size_t dot1 = ver.find_first_of('.',0);
+		size_t dot2 = ver.find_first_of('.',dot1 + 1);
+		size_t dot3 = ver.find_first_of('.',dot2 + 1);
 		if (dot3 == -1)  // since 5.0.7 patch is digit
 		{
 			dot3 = dot2 + 1;
-			while ( (dot3 < (int)ver.length()) && (isdigit(ver[dot3])) )
+			while ( (dot3 < ver.length()) && (isdigit(ver[dot3])) )
 			{
 				dot3++;
 			}
@@ -279,7 +279,7 @@ void Client::itemizeclientversion(const std::string& ver, VersionDetailStruct& d
 		detail.minor = atoi(ver.substr(dot1+1,dot2 - dot1 - 1).c_str());
 		detail.rev   = atoi(ver.substr(dot2+1,dot3 - dot2 - 1).c_str());
 		detail.patch = 0;
-		if (dot3<(int)ver.length())
+		if (dot3<ver.length())
 		{
 			if ( (detail.major<=5) && (detail.minor<=0) && (detail.rev<=6))
 			{
